# Move to new hash structure for even more performance