--- a/.gitignore
+++ b/.gitignore
@@ -2,3 +2,4 @@
 **/obj
 .ionide
 .vscode
+.fake