--- a/test/FSharp.HashCollections.Benchmarks/BenchmarkProgram.fs
+++ b/test/FSharp.HashCollections.Benchmarks/BenchmarkProgram.fs
@@ -4,6 +4,6 @@ open BenchmarkDotNet.Running
 
 [<EntryPoint>]
 let main argv =
-    let summary = BenchmarkRunner.Run(typeof<ReadBenchmarks>.Assembly)
+    let summary = BenchmarkSwitcher.FromAssembly(typeof<ReadBenchmarks>.Assembly).Run(argv)
     0