--- a/test/FSharp.HashCollections.Benchmarks/FSharp.HashCollections.Benchmarks.fsproj
+++ b/test/FSharp.HashCollections.Benchmarks/FSharp.HashCollections.Benchmarks.fsproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
 
@@ -28,6 +28,7 @@
     <PackageReference Include="FSharpX.Collections" Version="3.0.1" />
     <PackageReference Include="FSharpx.Collections.Experimental" Version="3.0.1" />
     <PackageReference Include="ImTools.dll" Version="1.0.0" />
+    <PackageReference Include="LanguageExt.FSharp" Version="4.2.9" />
     <PackageReference Include="System.Collections.Immutable" Version="1.7.0" />
   </ItemGroup>