… src/FSharp.HashCollections/HashMap.fs
… test/FSharp.HashCollections.Benchmarks/OfSeqBenchmark.fs
… test/FSharp.HashCollections.Benchmarks/ReadBenchmark.fs
… test/FSharp.HashCollections.Unit/HashMapTest.fs
… test/FSharp.HashCollections.Unit/HashSetTest.fs