This is an attempt to see the effect on performance if we remove nodes and bring the values up to avoid recursion for the entry node taking some inspiration from CHAMP but with a more .NET flavor taking advantage of some of F#'s features on top.

Current benchmarks below. For smaller collections below 1,000,000 significantly faster in many cases. As per comments this makes us the fastest in all cases across every collection in the .NET ecosystem I could find for the general lookup case. The very big collections grow slightly but not as much of a concern % wise. This is especially true around the 500,000 to 1,000,000 range.

Seeing 3.8GB go down to 2.1GB in memory consumption using a quick FSX script storing 30,000,000 elements with a HashMap<int, string> (string being converted from the int key) entry. So it is more memory efficient as well.

``` ini

BenchmarkDotNet=v0.13.1, OS=arch 
AMD Ryzen 7 3700X, 1 CPU, 16 logical and 8 physical cores
.NET SDK=6.0.110
  [Host]     : .NET 6.0.10 (6.0.1022.51301), X64 RyuJIT DEBUG
  DefaultJob : .NET 6.0.10 (6.0.1022.51301), X64 RyuJIT


```
|                           Method | CollectionSize |        Mean |    Error |   StdDev |
|--------------------------------- |--------------- |------------:|---------:|---------:|
|                       **GetHashMapCurrent** |             **10** |    **11.29 ns** | **0.092 ns** | **0.086 ns** |
|                       **GetHashMapNew** |             **10** |    **10.65 ns** | **0.058 ns** | **0.054 ns** |
|                       **GetHashMapCurrent** |            **100** |    **16.69 ns** | **0.135 ns** | **0.126 ns** |
|                       **GetHashMapNew** |            **100** |    **11.05 ns** | **0.066 ns** | **0.062 ns** |
|                       **GetHashMapCurrent** |           **1000** |    **12.53 ns** | **0.120 ns** | **0.101 ns** |
|                       **GetHashMapNew** |           **1000** |    **11.07 ns** | **0.071 ns** | **0.066 ns** |
|                       **GetHashMapCurrent** |         **100000** |    **27.60 ns** | **0.239 ns** | **0.223 ns** |
|                       **GetHashMapNew** |         **100000** |    **30.42 ns** | **0.274 ns** | **0.257 ns** |
|                       **GetHashMapCurrent** |         **500000** |    **75.84 ns** | **0.737 ns** | **0.689 ns** |
|                       **GetHashMapNew** |         **500000** |    **32.58 ns** | **0.302 ns** | **0.282 ns** |
|                       **GetHashMapCurrent** |         **750000** |   **108.36 ns** | **0.605 ns** | **0.506 ns** |
|                       **GetHashMapNew** |         **750000** |    **35.01 ns** | **0.387 ns** | **0.343 ns** |
|                       **GetHashMapCurrent** |        **1000000** |   **115.40 ns** | **0.202 ns** | **0.169 ns** |
|                       **GetHashMapNew** |        **1000000** |    **38.45 ns** | **0.111 ns** | **0.086 ns** |
|                       **GetHashMapCurrent** |        **5000000** |   **144.12 ns** | **0.331 ns** | **0.294 ns** |
|                       **GetHashMapNew** |        **5000000** |   **176.51 ns** | **0.731 ns** | **0.684 ns** |
|                       **GetHashMapCurrent** |       **10000000** |   **163.24 ns** | **0.788 ns** | **0.737 ns** |
|                       **GetHashMapNew** |       **10000000** |   **178.63 ns** | **0.561 ns** | **0.524 ns** |

``` ini